requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# Build-Bot/bbot.py: 12
psutil == 5.8.0

# Build-Bot/bbot.py: 13
pyfiglet == 0.8.post1

# Build-Bot/bbot.py: 17
python_telegram_bot == 13.2
